Added 'purge' kwarg to Model ctor to indicate if values no longer cached by subjects should be removed. Defaults to True.
When doing real-time plotting, you want to purge. Subjects have a
cache to determine the number of timetsamps worth of data (window
size) to use/display for analysis purposes. This correlates to the
number of rows in model produced dataframes which results in the
number of timestamps shown in time-series plots.

When building dataframes from Models populated from recorded data,
purging should not occur. In this case, you want to build a model
representative of all the recorded data so the resulting dataframe
will be complete.